@@ -7,7 +7,7 @@ import '../../../data/models/attendance_model.dart';
 import '../controllers/attendance_controller.dart';
 
 class AttendanceView extends GetView<AttendanceController> {
-  const AttendanceView({Key? key}) : super(key: key);
+  const AttendanceView({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -31,7 +31,7 @@ class AttendanceView extends GetView<AttendanceController> {
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
-              // Attendance Summary
+              
               AppCard(
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
@@ -81,7 +81,7 @@ class AttendanceView extends GetView<AttendanceController> {
               ),
               const SizedBox(height: 24),
               
-              // Attendance List
+              
               Text(
                 'Riwayat Kehadiran (30 Hari Terakhir)',
                 style: AppTextStyle.heading3,
